Mise en forme + suppression d'un truc en double

+++ src/fr/ecole3il/rodez2023/perlin/math/BruitPerlin2D.java
@@ -37,7 +37,7 @@ public class BruitPerlin2D extends Bruit2D {
 	public double bruit2D(double x, double y) {
 		double tempX, tempY;
 		int x0, y0, ii, jj, gi0, gi1, gi2, gi3;
-		double unit = 1.0f / (double) Math.sqrt(2);
+		double unit = 1.0f / Math.sqrt(2);
 		double tmp, s, t, u, v, Cx, Cy, Li1, Li2;
 		// Adapter pour la résolution
 		x /= getResolution();
